// src/utils/gridViewPayLoad.d.ts
export interface gridViewPayLoadSupplier{
    gridState?: {
        columns?: [
            {
                visible?: boolean,
                width?: string,
                name?: string,
                sort?:{
                    direction: string
                }
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
                filters?: {
                    filters?: [
                        {
                            field?: string,
                            operator?: string,
                            value?: Number
                        }
                    ],
                    logic?: string
                }
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            }
        ],
        scrollFocus?: {},
        selection?: [],
        grouping?: {},
        treeView?: {},
        pagination?:{
            paginationCurrentPage?: Number,
            paginationPageSize?: Number
        },
        filter?: {
            filters?: [
                {
                    filters?:[
                        {
                            field?: string,
                            operator?: string,
                            value?: Number
                        }
                    ],
                    logic?: string
                }
            ],
            logic?: string
        },
        sort?: [
            {
                dir?: string,
                field?: string,
            }
        ]
    }
    name?: string
    itemType?: string 
};